refactor[#3] : category directory 위치 변경
